feat(course): add Student

+++ src/main/java/nextstep/courses/domain/Session.java
@@ -5,7 +5,7 @@ import nextstep.users.domain.NsUser;
 import java.time.LocalDate;
 
 public class Session {
-    private final int id;
+    private final Long id;
 
     private final  SessionInfo sessionInfo;
 
@@ -19,14 +19,14 @@ public class Session {
                    ChargeStatus chargeStatus, int price,
                    int capacity, int sizeOfStudents, SessionStatus sessionStatus,
                    LocalDate startDate, LocalDate endDate) {
-        this(0,
+        this(0L,
                 new SessionInfo(title, creatorId, coverImage),
                 new Charge(chargeStatus, price),
                 new Enrollment(capacity, sizeOfStudents, sessionStatus),
                 new SessionPeriod(startDate, endDate));
     }
 
-    public Session(int id, String title, Long creatorId, String coverImage,
+    public Session(Long id, String title, Long creatorId, String coverImage,
                    ChargeStatus chargeStatus, int price,
                    int capacity, int sizeOfStudents, SessionStatus sessionStatus,
                    LocalDate startDate, LocalDate endDate) {
@@ -37,7 +37,7 @@ public class Session {
                 new SessionPeriod(startDate, endDate));
     }
 
-    public Session(int id, SessionInfo sessionInfo, Charge charge, Enrollment enrollment, SessionPeriod sessionPeriod) {
+    public Session(Long id, SessionInfo sessionInfo, Charge charge, Enrollment enrollment, SessionPeriod sessionPeriod) {
         this.id = id;
         this.sessionInfo = sessionInfo;
         this.charge = charge;
